requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=9.4.0
numpy>=1.24.0
flask>=2.3.0
ultralytics>=8.0.0
matplotlib>=3.7.0
scikit-image>=0.20.0
hydra-core>=1.3.0
iopath>=0.1.10
tqdm>=4.66.0 